backend/requirements.txt
# Web framework
fastapi==0.116.0
uvicorn[standard]==0.35.0

# Multipart form parsing & env vars
python-multipart==0.0.20
python-dotenv==1.1.1

# PDF/DOCX parsing
PyMuPDF==1.26.3
docx2txt==0.9

# Google Cloud & Firebase (firebase-admin needs storage>=1.37.1)
google-cloud-storage>=1.37.1
firebase-admin==6.3.0
google-auth==2.40.3

# Google Generative AI client
google-generativeai==0.8.5
